The database has been changed, the annotation in entities has been changed

+++ src/main/java/com/accenture/russiaatc/irentservice/access/ParkingRepository.java
@@ -3,5 +3,8 @@ package com.accenture.russiaatc.irentservice.access;
 import com.accenture.russiaatc.irentservice.model.entity.Parking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParkingRepository extends JpaRepository<Parking, Integer>  {
+import java.util.Optional;
+
+public interface ParkingRepository extends JpaRepository<Parking, Long>  {
+
 }
