+++ src/main/java/com/accenture/russiaatc/irentservice/access/RentRepository.java
@@ -3,5 +3,5 @@ package com.accenture.russiaatc.irentservice.access;
 import com.accenture.russiaatc.irentservice.model.entity.Rent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentRepository extends JpaRepository <Rent, Integer> {
+public interface RentRepository extends JpaRepository <Rent, Long> {
 }
